requirements.txt
absl-py==0.9.0
cachetools==4.1.1
chardet==3.0.4
cloudpickle==1.2.1
decorator==4.4.2
dgl==0.3
future==0.17.1
google-auth==1.20.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
gym==0.15.4
-e git+https://github.com/agakshat/gym_vecenv.git@1d2dfd260626525c76295924f34fb1d9082186e9#egg=gym_vecenv
idna==2.8
importlib-metadata==1.7.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Markdown==3.2.2
matplotlib==2.2.3
networkx==2.5
numpy==1.15.1
numpy-stl==2.10.0
oauthlib==3.1.0
pandas==0.23.4
pickleshare==0.7.4
Pillow==5.2.0
protobuf==3.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pyglet==1.3.2
python-utils==2.3.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.21.3
scipy==1.2.1
six==1.12.0
seaborn==0.9.0
stable-baselines==2.8.0
tensorboard==1.14.0
tensorboardX==1.6
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torchvision==0.2.1
urllib3==1.25.3
Werkzeug==1.0.1
xgboost==1.1.1
zipp==3.1.0
